requirements.txt
aiohttp
fastapi
openai
pydantic
python-dotenv
python-telegram-bot
redis
uvicorn